# ns-topup/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NS Games Topup API",
        "description": "API untuk platform topup game, pulsa, dan e-wallet dengan integrasi Digiflazz dan Midtrans",
        "contact": {
            "name": "NS Games Support",
            "email": "admin@nsgames.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Development Server"
        },
        {
            "url": "http://localhost/NS-topupgames/ns-topup/public",
            "description": "Laragon/XAMPP Server"
        }
    ],
    "paths": {
...
        "/api/digiflazz/balance": {
            "get": {
                "tags": [
                    "Digiflazz"
                ],
                "summary": "Get Digiflazz balance",
                "operationId": "37f5a09dfbc0daa605c8bd8cdf4f130e",
                "responses": {
                    "200": {
                        "description": "Balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "balance": {
                                            "type": "number",
                                            "example": 100000
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/digiflazz/price-list": {
            "get": {
                "tags": [
                    "Digiflazz"
                ],
                "summary": "Get Digiflazz price list",
                "operationId": "b329044644ee4c7df019913e6b749274",
                "responses": {
                    "200": {
                        "description": "Price list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product_name": {
                                                                "type": "string",
                                                                "example": "Mobile Legends 86 Diamond"
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "Games"
                                                            },
                                                            "brand": {
                                                                "type": "string",
                                                                "example": "MOBILE LEGENDS"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "example": 20000
                                                            },
                                                            "buyer_sku_code": {
                                                                "type": "string",
                                                                "example": "ML86"
                                                            },
                                                            "buyer_product_status": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "stock": {
                                                                "type": "number",
                                                                "example": 999999
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "sku_code": {
                                                    "type": "string",
                                                    "example": "ML86"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "86 Diamond"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Game"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "digiflazz"
                                                },
                                                "game": {
                                                    "type": "object"
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 86
                                                },
                                                "base_price": {
                                                    "type": "number",
                                                    "example": 20000
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": 22000
                                                },
                                                "profit": {
                                                    "type": "number",
                                                    "example": 2000
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": -1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}